# More efficient parsing of JSON arrays and objects